Refactor export APIs related code
